refactor: move user enums to its module
