+++ packages/apps/reputation-oracle/server/src/common/decorators/index.ts
@@ -1,8 +1,8 @@
 import { SetMetadata } from '@nestjs/common';
-import { Role } from '../enums/user';
+import { UserRole } from '../../modules/user';
 
 export const Public = (): ((target: any, key?: any, descriptor?: any) => any) =>
   SetMetadata('isPublic', true);
 
 export const ROLES_KEY = 'roles';
-export const Roles = (...roles: Role[]) => SetMetadata(ROLES_KEY, roles);
+export const Roles = (...roles: UserRole[]) => SetMetadata(ROLES_KEY, roles);
