--- packages/apps/reputation-oracle/server/src/common/interfaces/user.ts
@@ -1,9 +1,0 @@
-import { UserStatus, Role } from '../enums/user';
-import { IBase } from './base';
-
-export interface IUser extends IBase {
-  password: string;
-  email: string;
-  status: UserStatus;
-  role: Role;
-}
