+++ packages/apps/reputation-oracle/server/src/modules/user/index.ts
@@ -1,0 +1,4 @@
+export { UserEntity, UserStatus, Role as UserRole } from './user.entity';
+export { UserRepository } from './user.repository';
+export { UserService } from './user.service';
+export { UserModule } from './user.module';
