+++ packages/apps/reputation-oracle/server/src/modules/user/user.service.ts
@@ -3,12 +3,7 @@ import { Injectable } from '@nestjs/common';
 import * as bcrypt from 'bcrypt';
 import { HCaptchaConfigService } from '../../config/hcaptcha-config.service';
 import { Web3ConfigService } from '../../config/web3-config.service';
-import {
-  KycStatus,
-  OperatorStatus,
-  UserStatus,
-  Role,
-} from '../../common/enums/user';
+import { KycStatus, OperatorStatus } from '../../common/enums/user';
 import { SignatureType } from '../../common/enums/web3';
 import { SiteKeyType } from '../../common/enums';
 import { HCaptchaService } from '../../integrations/hcaptcha/hcaptcha.service';
@@ -21,7 +16,7 @@ import { Web3Service } from '../web3/web3.service';
 import { SiteKeyEntity } from './site-key.entity';
 import { SiteKeyRepository } from './site-key.repository';
 import { RegisterAddressRequestDto } from './user.dto';
-import { UserEntity } from './user.entity';
+import { Role, UserStatus, UserEntity } from './user.entity';
 import {
   UserError,
   UserErrorMessage,
